[task] new folder for cpp files
