--- code/comment.cpp
@@ -1,11 +1,0 @@
-// this is a single line comment
-
-/* But this is 
-a plenty of lines
-comment*/
-
-/////////////////
-// This is a   //
-// comment box //
-/////////////////
-
