+++ code/cpp_sources/comment.cpp
@@ -1,0 +1,11 @@
+// this is a single line comment
+
+/* But this is 
+a plenty of lines
+comment*/
+
+/////////////////
+// This is a   //
+// comment box //
+/////////////////
+
